[Refactor] change project name to Earth1Round
