+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'donggrami-server'
+rootProject.name = 'earth1round'
