+++ src/main/java/donggrami/earth1round/Earth1RoundApplication.java
@@ -1,13 +1,13 @@
-package donggrami.donggramiserver;
+package donggrami.earth1round;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DonggramiServerApplication {
+public class Earth1RoundApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DonggramiServerApplication.class, args);
+		SpringApplication.run(Earth1RoundApplication.class, args);
 	}
 
 }
