+++ src/main/java/donggrami/earth1round/config/BaseException.java
@@ -1,4 +1,4 @@
-package donggrami.donggramiserver.config;
+package donggrami.earth1round.config;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
