+++ src/main/java/donggrami/earth1round/config/BaseResponse.java
@@ -1,4 +1,4 @@
-package donggrami.donggramiserver.config;
+package donggrami.earth1round.config;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import static donggrami.donggramiserver.config.BaseResponseStatus.SUCCESS;
+import static donggrami.earth1round.config.BaseResponseStatus.SUCCESS;
 
 @Getter
 @AllArgsConstructor
