+++ src/test/java/donggrami/earth1round/Earth1RoundApplicationTests.java
@@ -1,10 +1,10 @@
-package donggrami.donggramiserver;
+package donggrami.earth1round;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DonggramiServerApplicationTests {
+class Earth1RoundApplicationTests {
 
 	@Test
 	void contextLoads() {
